+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # alpinepass
 
 ## UNRELEASED
+
+* New command "out" for exporting system information to a file.
+* New command "show" for printing system information to the console.
+* New "passwords" flag for including passwords in the output.